+++ contracts/staking/test/utils/queue.ts
@@ -1,7 +1,9 @@
+import * as _ from 'lodash';
+
 export class Queue<T> {
     private _store: T[] = [];
     constructor (store?: T[]) {
-        this._store = store !== undefined ? store : [];
+        this._store = store !== undefined ? _.cloneDeep(store) : [];
     }
     public pushBack(val: T): void {
         this._store.push(val);
